Fix some tests. Move integration tests to  folder
